[builder] create generic unparent object method
Currently wheels of a wheel_robot are removed by builder before starting the game engine.  This is useful for other objects connected with physics constraints, not just wheeled robots.  A generic unparent_objects method was added to abstract component to allow a list of child objects to be removed.  The WheeledRobot class then uses this method to remove the wheels as children